src/koth/views.py
from django.shortcuts import render
from koth.models import Box, Player, Game, GamePlayer
from koth.serializers import BoxSerializer, PlayerSerializer, GameSerializer, GamePlayerSerializer
from rest_framework import viewsets

class BoxViewSet(viewsets.ModelViewSet):
    queryset = Box.objects.all()
    serializer_class = BoxSerializer

class PlayerViewSet(viewsets.ModelViewSet):
    queryset = Player.objects.all()
    serializer_class = PlayerSerializer

class GameViewSet(viewsets.ModelViewSet):
    queryset = Game.objects.all()
    serializer_class = GameSerializer

class GamePlayerViewSet(viewsets.ModelViewSet):
    queryset = GamePlayer.objects.all()
    serializer_class = GamePlayerSerializer
